Rename type FeatureColumnMap into ColumnMap
